+++ ihp/Test/Main.hs
@@ -1,5 +1,7 @@
 module Main where
 
+import IHP.Prelude
+
 import Test.Hspec
 import qualified AutoRefreshSpec